requirements.txt
streamlit
pandas
scikit-learn
pillow
numpy
pytesseract
